+++ Craftsman/Enums/ExchangeType.cs
@@ -6,6 +6,6 @@
         Direct,
         Fanout,
         Topic,
-        Header
+        //Header
     }
 }