+++ CHANGELOG.md
@@ -36,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Bus, Messages, Producers, and Consumers props to BC template
 - Added `add:message` command
 - Updated entity name and entity prop names first letter to always be capitalized
-- Added `add:consumer` and `add:producer` commands
+- Added `add:consumer` and `add:producer` commands for direct, topic, and fanout messages
 - Removed BC readme and updated sln readme
 - Better namespacing for features in controllers using static classes for features
 - Updated functional test to pass without conflict